Resolve the import conflict during package execution
